Fix custom attrs flattening across widget interfaces
