+++ packages/user-management-widget/src/lib/helpers.ts
@@ -1,0 +1,28 @@
+import { User } from './widget/api/types';
+
+const unflattenKeys = ['customAttributes'];
+
+export const unflatten = (formData: Partial<User>) =>
+  Object.entries(formData).reduce((acc, [key, value]) => {
+    const [prefix, ...rest] = key.split('.');
+
+    if (!unflattenKeys.includes(prefix)) {
+      return Object.assign(acc, { [key]: value });
+    }
+
+    if (!acc[prefix]) {
+      acc[prefix] = {};
+    }
+
+    acc[prefix][rest.join('.')] = value;
+
+    return acc;
+  }, {});
+
+export const flatten = (
+  vals: Record<string, string | boolean | number>,
+  keyPrefix: string,
+) =>
+  Object.fromEntries(
+    Object.entries(vals || {}).map(([key, val]) => [`${keyPrefix}${key}`, val]),
+  );
