+++ packages/user-management-widget/src/lib/widget/mixins/initMixin/initComponentsMixins/initEditUserModalMixin.ts
@@ -11,7 +11,6 @@ import {
 } from '@descope/sdk-helpers';
 import { formMixin, loggerMixin, modalMixin } from '@descope/sdk-mixins';
 import parsePhone from 'libphonenumber-js/min';
-import { User } from '../../../api/types';
 import { stateManagementMixin } from '../../stateManagementMixin';
 import { initWidgetRootMixin } from './initWidgetRootMixin';
 import {
@@ -19,33 +18,7 @@ import {
   getSelectedUsers,
   getTenantRoles,
 } from '../../../state/selectors';
-
-const unflattenKeys = ['customAttributes'];
-
-const unflatten = (formData: Partial<User>) =>
-  Object.entries(formData).reduce((acc, [key, value]) => {
-    const [prefix, ...rest] = key.split('.');
-
-    if (!unflattenKeys.includes(prefix)) {
-      return Object.assign(acc, { [key]: value });
-    }
-
-    if (!acc[prefix]) {
-      acc[prefix] = {};
-    }
-
-    acc[prefix][rest.join('.')] = value;
-
-    return acc;
-  }, {});
-
-const flatten = (
-  vals: Record<string, string | boolean | number>,
-  keyPrefix: string,
-) =>
-  Object.fromEntries(
-    Object.entries(vals || {}).map(([key, val]) => [`${keyPrefix}${key}`, val]),
-  );
+import { flatten, unflatten } from '../../../../helpers';
 
 const formatPhoneNumber = (phoneNumber: string) => {
   if (!phoneNumber) return phoneNumber;
